chore: move is* functions to common
